Reduce complexity of EvaluationRule
Creates rules.py containing logic for the small rule snippets for each
entry in the json
